requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
pytz>=2024.1
python-dotenv>=1.0.0
apscheduler>=3.10.4
tweepy>=4.14.0
